chore: folder restructure
modifying name of the /components/commons folder to /components/global
given that it's more descrptive
