+++ src/components/constants/web_strings.ts
@@ -1,3 +1,11 @@
 // String contants
 
 export const contentDescription = 'A technical assignment submitted to Savannah Informatics by Zidane Gimiga';
+
+//navbar
+export const navBar = {
+    home: ["Home", "/"],    
+    users: ["Users", "/users"],    
+    posts: ["Posts", "/posts"],    
+    about: ["About Project", "www.github.com/zidanegimiga/react-interview-project"]    
+}
