+++ src/components/global/Navbar.tsx
@@ -1,0 +1,37 @@
+import Link from "next/link";
+import MobileNavbar from "./MobileNavbar";
+import styles from "../../../styles/Navbar.module.scss";
+
+const Navbar = () => {
+  return (
+    <header className={styles.nav}>
+      <MobileNavbar />
+      <section className={styles.mainWrapper}>
+        <div>
+          <Link href="/" passHref>
+            <a className={styles.homeLink}>Home</a>
+          </Link>
+        </div>
+        <div className={styles.rightSection}>
+          <div className={styles.linksRightContainer}>
+            <Link href="/users" passHref>
+              <a className={styles.rightSideLink}> Users</a>
+            </Link>
+            <Link href="/posts" passHref>
+              <a className={styles.rightSideLink}>Posts</a>
+            </Link>
+            <a
+              href="www.github.com/zidanegimiga"
+              target="_blank"
+              className={styles.rightSideLink}
+            >
+              About Project
+            </a>
+          </div>
+        </div>
+      </section>
+    </header>
+  );
+};
+
+export default Navbar;
